working version for Zee in diphoton trigger

+++ Taggers/python/flashggTagSorter_cfi.py
@@ -24,7 +24,7 @@ flashggTagSorter = cms.EDProducer('FlashggTagSorter',
         ###                                                                 cms.PSet(TagName = cms.InputTag('flashggSigmaMoMpToMTag'))
         ),
                                   MassCutUpper=cms.double(180.),
-                                  MassCutLower=cms.double(100),
+                                  MassCutLower=cms.double(60),
                                   MinObjectWeightException = cms.double(0.0),
                                   MaxObjectWeightException = cms.double(10.),
                                   MinObjectWeightWarning = cms.double(0.4),
